requirements.txt
Django==4.2.10
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
channels==4.0.0
daphne==4.0.0
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
django-cors-headers==4.3.1
django-filter==23.5
python-decouple==3.8
drf-yasg==1.21.7
pillow==10.2.0
markdown==3.5.2
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
dj-database-url==2.1.0 